+++ pip_check_reqs/find_extra_reqs.py
@@ -13,8 +13,11 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Callable, Iterable
 from unittest import mock
 
-from packaging.utils import canonicalize_name
-from pip._internal.commands.show import _PackageInfo, search_packages_info
+from packaging.utils import NormalizedName, canonicalize_name
+from pip._internal.commands.show import (
+    _PackageInfo,  # pyright: ignore[reportPrivateUsage]
+    search_packages_info,
+)
 
 from pip_check_reqs import common
 from pip_check_reqs.common import version_info
@@ -24,6 +27,23 @@ if TYPE_CHECKING:
 
 log = logging.getLogger(__name__)
 
+@cache
+def get_packages_info() -> list[_PackageInfo]:
+    all_pkgs = [
+        dist.metadata["Name"] for dist in importlib.metadata.distributions()
+    ]
+
+    # On Python 3.11 (and maybe higher), setting this environment variable
+    # dramatically improves speeds.
+    # See https://github.com/r1chardj0n3s/pip-check-reqs/issues/123.
+    with mock.patch.dict(os.environ, {"_PIP_USE_IMPORTLIB_METADATA": "False"}):
+        return list(search_packages_info(query=all_pkgs))
+
+
+# This is a slow operation.
+# It only happens once when calling the CLI, but it is hit many times in
+# tests.
+# We cache the result to speed up tests.
 @cache
 def get_packages_info() -> list[_PackageInfo]:
     all_pkgs = [
@@ -57,18 +77,8 @@ def find_extra_reqs(
         ignore_modules_function=ignore_modules_function,
     )
 
-    # 2. find which packages provide which files
-    installed_files = {}
-    all_pkgs = [
-        dist.metadata["Name"] for dist in importlib.metadata.distributions()
-    ]
-
-    # On Python 3.11 (and maybe higher), setting this environment variable
-    # dramatically improves speeds.
-    # See https://github.com/r1chardj0n3s/pip-check-reqs/issues/123.
-    with mock.patch.dict(os.environ, {"_PIP_USE_IMPORTLIB_METADATA": "False"}):
-        packages_info = list(search_packages_info(all_pkgs))
-
+    installed_files: dict[Path, str] = {}
+    packages_info = get_packages_info()
     here = Path().resolve()
 
     for package in packages_info:
@@ -93,9 +103,9 @@ def find_extra_reqs(
             package_location,
         )
         for package_file in package_files:
-            path = str(
-                (Path(package_location) / package_file).resolve(),
-            )
+            path = Path(package_location) / package_file
+            path = path.resolve()
+
             installed_files[path] = package_name
             package_path = common.package_path(path=path)
             if package_path:
@@ -105,7 +115,10 @@ def find_extra_reqs(
                 installed_files[package_path] = package_name
 
     # 3. match imported modules against those packages
-    used = collections.defaultdict(list)
+    used: collections.defaultdict[
+        NormalizedName,
+        list[common.FoundModule],
+    ] = collections.defaultdict(list)
 
     for modname, info in used_modules.items():
         # probably standard library if it's not in the files list